+++ src/test/java/de/samply/core/ExporterCoreTest.java
@@ -50,7 +50,7 @@ class ExporterCoreTest {
     @Test
     void retrieveByQueryId() throws ExporterCoreException {
         ExporterParameters exporterParameters = new ExporterParameters(null, "Patient",
-                converterTemplateId, null, null, Format.FHIR_QUERY, null, null, null, null, Format.CSV);
+                converterTemplateId, null, null, Format.FHIR_PATH, null, null, null, null, Format.CSV);
         ExporterCoreParameters exporterCoreParameters = exporterCore.extractParameters(
                 exporterParameters);
         Flux<Path> resultFlux = exporterCore.retrieveQuery(exporterCoreParameters);